+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.10.0
 
 - In BSearch its possible to hide the reset button
+- Redesign BAlert <https://github.com/abaldeweg/ui/blob/main/packages/%40baldeweg/ui/docu/components.md#balert>
+- 3 new CSS vars for the blue info alerts
 
 ## v0.9.0